refactor legacy CSV importer into modular components for improved maintainability
