+++ backend/apps/imports/csv_importer/legacy_csv_relations.py
@@ -1,0 +1,69 @@
+"""Model relation helpers for legacy CSV imports."""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+from django.utils.text import slugify
+
+from apps.genres.models import Genre, GenreTranslation, GenreUseAs
+from apps.languages.models import Language
+from apps.locations.models import Hall, HallTranslation
+from apps.productions.models import ProductionGenre
+
+from .legacy_csv_constants import LEGACY_LANGUAGE_ACTIVE, LEGACY_LANGUAGE_CODE, LEGACY_LANGUAGE_NAME
+
+if TYPE_CHECKING:
+    from apps.productions.models import Production
+
+
+def _ensure_language() -> Language:
+    language, _ = Language.objects.get_or_create(
+        code=LEGACY_LANGUAGE_CODE,
+        defaults={"name": LEGACY_LANGUAGE_NAME, "is_active": LEGACY_LANGUAGE_ACTIVE},
+    )
+    return language
+
+
+def _ensure_genre(label: str, language: Language) -> Genre:
+    genre_use_as, _ = GenreUseAs.objects.get_or_create(name="genre")
+    genre_type = slugify(label)[:50] or label.strip().lower().replace(" ", "_")[:50] or "genre"
+
+    translation = (
+        GenreTranslation.objects.select_related("genre")
+        .filter(language=language, name=label[:50], genre__use_as=genre_use_as)
+        .order_by("genre_id")
+        .first()
+    )
+    if translation:
+        return translation.genre
+
+    genre = Genre.objects.filter(type=genre_type, use_as=genre_use_as).order_by("id").first()
+    if genre is None:
+        genre = Genre.objects.create(type=genre_type, use_as=genre_use_as)
+
+    GenreTranslation.objects.update_or_create(genre=genre, language=language, defaults={"name": label[:50]})
+    return genre
+
+
+def _sync_production_genres(production: Production, genre_labels: list[str], language: Language) -> None:
+    ProductionGenre.objects.filter(production=production).delete()
+    links = []
+    for position, label in enumerate(genre_labels):
+        genre = _ensure_genre(label, language)
+        links.append(ProductionGenre(production=production, genre=genre, position=position))
+    if links:
+        ProductionGenre.objects.bulk_create(links)
+
+
+def _hall_for_name(name: str, language: Language) -> Hall | None:
+    if not name:
+        return None
+
+    translation = HallTranslation.objects.select_related("hall").filter(name=name).first()
+    if translation:
+        return translation.hall
+
+    hall = Hall.objects.create(space=None, seat_selection=False, open_seating=False)
+    HallTranslation.objects.create(hall=hall, language=language, name=name)
+    return hall
